feat: add function call support to CactusLM, improve serialization, and standardize quantization to int4 across integration tests.

+++ example/integration_test/audio_speaker_embed_test.dart
@@ -29,7 +29,7 @@ void main() {
     setUpAll(() async {
       audio = CactusAudio(
         model: _spkModel,
-        options: const CactusModelOptions(quantization: 'int8'),
+        options: const CactusModelOptions(quantization: 'int4'),
       );
       pcmData = await _loadPcmFromAsset();
     });
